move component loaders to source files
