+++ headers/enduro2d/high/components/actor.hpp
@@ -8,6 +8,7 @@
 
 #include "../_high.hpp"
 
+#include "../component.hpp"
 #include "../node.hpp"
 
 namespace e2d
@@ -23,6 +24,20 @@ namespace e2d
         const_node_iptr node() const noexcept;
     private:
         node_iptr node_;
+    };
+
+    template <>
+    class component_loader<actor> {
+    public:
+        static const char* schema_source;
+
+        bool operator()(
+            actor& component,
+            const component_loader<>::fill_context& ctx) const;
+            
+        bool operator()(
+            asset_dependencies& dependencies,
+            const component_loader<>::collect_context& ctx) const;
     };
 }
