+++ headers/enduro2d/high/components/flipbook_source.hpp
@@ -8,6 +8,7 @@
 
 #include "../_high.hpp"
 
+#include "../component.hpp"
 #include "../assets/flipbook_asset.hpp"
 
 namespace e2d
@@ -21,6 +22,20 @@ namespace e2d
         const flipbook_asset::ptr& flipbook() const noexcept;
     private:
         flipbook_asset::ptr flipbook_;
+    };
+
+    template <>
+    class component_loader<flipbook_source> {
+    public:
+        static const char* schema_source;
+
+        bool operator()(
+            flipbook_source& component,
+            const component_loader<>::fill_context& ctx) const;
+            
+        bool operator()(
+            asset_dependencies& dependencies,
+            const component_loader<>::collect_context& ctx) const;
     };
 }
