+++ headers/enduro2d/high/components/model_renderer.hpp
@@ -8,6 +8,7 @@
 
 #include "../_high.hpp"
 
+#include "../component.hpp"
 #include "../assets/model_asset.hpp"
 
 namespace e2d
@@ -21,6 +22,20 @@ namespace e2d
         const model_asset::ptr& model() const noexcept;
     private:
         model_asset::ptr model_;
+    };
+
+    template <>
+    class component_loader<model_renderer> {
+    public:
+        static const char* schema_source;
+
+        bool operator()(
+            model_renderer& component,
+            const component_loader<>::fill_context& ctx) const;
+
+        bool operator()(
+            asset_dependencies& dependencies,
+            const component_loader<>::collect_context& ctx) const;
     };
 }
